pyproject.toml
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "texta-ai"
version = "1.0.0"
description = "Assistente de Correção de Texto Inteligente usando IA"
authors = ["Texta AI Team <your-email@example.com>"] # Adicione um e-mail para o Poetry
readme = "README.md"
packages = [{include = "src"}] # Assumindo que o código principal está em 'src/'
keywords = ["ai", "text-correction", "productivity", "automation"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Office/Business :: Office Suites",
    "Topic :: Text Processing :: Linguistic",
]

[tool.poetry.dependencies]
python = ">=3.9,<3.13" # Especificação de versão para Poetry
openai-agents = "^0.0.11"
openai = "^1.23.6"
pyperclip = "^1.8.2"
python-dotenv = "^1.0.1"
pyautogui = "^0.9.54"
PySide6 = "^6.7.0"
pynput = { version = ">=1.7.6" } # Formato para especificar versão mínima
loguru = "^0.7.2"

[tool.poetry.group.dev.dependencies]
pytest = "^8.1.1"
pytest-mock = "^3.14.0"
pytest-cov = "^5.0.0"
pytest-asyncio = "^0.23.6"
ruff = { version = ">=0.8.0" }
# Black e isort são removidos, pois Ruff agora lida com formatação e ordenação de imports.
# black = { version = ">=24.0.0" }
# isort = { version = ">=5.13.0" }

# =============================================================================
# RUFF CONFIGURATION
# =============================================================================
[tool.ruff]
# Configurações gerais
line-length = 88
indent-width = 4
target-version = "py39"

# Arquivos e diretórios a serem excluídos
exclude = [
    ".bzr",
    ".direnv", 
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".ipynb_checkpoints",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pyenv",
    ".pytest_cache",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    ".vscode",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "site-packages",
    "venv",
    "logs",
]

[tool.ruff.lint]
# Regras habilitadas - configuração otimizada para projetos Python
select = [
    "E",     # Pycodestyle errors
    "W",     # Pycodestyle warnings
    "F",     # Pyflakes
    "I",     # Import sorting (isort)
    "N",     # Naming conventions (PEP8)
    "D",     # Docstring style (pydocstyle)
    "UP",    # Pyupgrade - modernização do código
    "YTT",   # flake8-2020
    "ANN",   # flake8-annotations
    "S",     # flake8-bandit (segurança)
    "BLE",   # flake8-blind-except
    "FBT",   # flake8-boolean-trap
    "B",     # flake8-bugbear
    "A",     # flake8-builtins
    "COM",   # flake8-commas
    "C4",    # flake8-comprehensions
    "DTZ",   # flake8-datetimez
    "T10",   # flake8-debugger
    "EM",    # flake8-errmsg
    "EXE",   # flake8-executable
    "FA",    # flake8-future-annotations
    "ISC",   # flake8-implicit-str-concat
    "ICN",   # flake8-import-conventions
    "G",     # flake8-logging-format
    "INP",   # flake8-no-pep420
    "PIE",   # flake8-pie
    "T20",   # flake8-print
    "PYI",   # flake8-pyi
    "PT",    # flake8-pytest-style
    "Q",     # flake8-quotes
    "RSE",   # flake8-raise
    "RET",   # flake8-return
    "SLF",   # flake8-self
    "SLOT",  # flake8-slots
    "SIM",   # flake8-simplify
    "TID",   # flake8-tidy-imports
    "TCH",   # flake8-type-checking
    "INT",   # flake8-gettext
    "ARG",   # flake8-unused-arguments
    "PTH",   # flake8-use-pathlib
    "ERA",   # eradicate (comentários mortos)
    "PD",    # pandas-vet
    "PGH",   # pygrep-hooks
    "PL",    # Pylint
    "TRY",   # tryceratops
    "FLY",   # flynt
    "NPY",   # NumPy specific rules
    "PERF",  # Perflint
    "RUF",   # Ruff-specific rules
]

# Regras ignoradas (ajuste conforme necessário)
ignore = [
    "E501",   # Line too long (gerenciado pelo formatter)
    "D100",   # Missing docstring in public module
    "D101",   # Missing docstring in public class 
    "D102",   # Missing docstring in public method
    "D103",   # Missing docstring in public function
    "D104",   # Missing docstring in public package
    "D105",   # Missing docstring in magic method
    "D107",   # Missing docstring in __init__
    "S101",   # Use of assert detected
    "COM812", # Trailing comma missing (conflita com formatter)
    "COM819", # Trailing comma prohibited (conflita com formatter)
    "ISC001", # Implicitly concatenated string literals (conflita com formatter)
    "Q000",   # Single quotes found but double quotes preferred (conflita com formatter)
    "Q001",   # Single quote docstring found but double quotes preferred
    "Q002",   # Double quote docstring found but single quotes preferred
    "Q003",   # Change outer quotes to avoid escaping inner quotes
]

# Permite correção automática para todas as regras habilitadas
fixable = ["ALL"]
unfixable = []

# Permite variáveis não utilizadas quando prefixadas com underscore
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.lint.per-file-ignores]
# Configurações específicas por arquivo
"__init__.py" = ["E402", "F401", "F403", "F811"] # Imports em __init__.py
"tests/**/*" = ["S101", "PLR2004", "ANN", "D"] # Testes podem usar assert, números mágicos, etc.
"scripts/**/*" = ["T20", "S101", "N999"] # Scripts podem usar print, assert e nomes com hífens

[tool.ruff.lint.pydocstyle]
# Convenção do Google para docstrings
convention = "google"

[tool.ruff.lint.pylint]
# Configurações do Pylint
max-args = 7
max-branches = 15
max-returns = 8
max-statements = 60

[tool.ruff.lint.isort]
# Configurações do isort integradas no Ruff
known-first-party = ["src"] # Informa ao Ruff que 'src' é código do projeto
force-sort-within-sections = true
split-on-trailing-comma = true

[tool.ruff.format]
# Configurações do formatter (compatível com Black)
quote-style = "double"
indent-style = "space"

# =============================================================================
# PYTEST CONFIGURATION
# =============================================================================
[tool.pytest.ini_options]
# Configuração do pytest (migrada de pytest.ini)
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--strict-markers",
    "--strict-config", 
    "--verbose",
    "--tb=short",
    "--cov=src",
    "--cov-report=term-missing",
    "--cov-report=html:htmlcov",
    "--cov-report=xml",
    "--cov-fail-under=65",
]
asyncio_default_fixture_loop_scope = "function"
asyncio_mode = "strict"
filterwarnings = [
    "ignore::pytest.PytestUnraisableExceptionWarning",
    "ignore::RuntimeWarning:asyncio",
    "ignore:coroutine.*was never awaited:RuntimeWarning",
]
cache_dir = ".pytest_temp_cache"
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
]

# =============================================================================
# COVERAGE CONFIGURATION
# =============================================================================
[tool.coverage.run]
source = ["src"]
omit = [
    "*/tests/*",
    "*/test_*",
    "*/__pycache__/*",
    "*/site-packages/*",
    "setup.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "if settings.DEBUG",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if __name__ == .__main__.:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]
show_missing = true
precision = 2

[tool.coverage.html]
directory = "htmlcov" 